add galleryProduct upload

+++ OOOFormula/Services/FilesRepository.cs
@@ -41,21 +41,27 @@ namespace OOOFormula.Services
         public void DeleteFile(string Image, string Folder)
         {
             string filePath = Path.Combine(_webHostEnvironment.WebRootPath, "images", Folder, Image); //создаем полное имя файла
-            if (Image != "noimage.png") //проверяем, не используется ли заглушка
+            if (Image != "noimage.png")//проверяем, не используется ли заглушка
             {
-                File.Delete(filePath);
+                if (File.Exists(filePath)) //проверяем существует ли файл
+                {
+                    File.Delete(filePath);
+                }
             }
         }
 
-        public async Task<string> UploadFile(IFormFile photo, string Folder)
+        public async Task<string> UploadFile(IFormFile photo, string Folder, string subFolder = null)
         {
             string uniqueFileName = null;
             if (photo != null)
             {
-                string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath, "images", Folder); //webRootPath возвращает путь до каталогаа wwwroot
+                string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath, "images", Folder, subFolder); //webRootPath возвращает путь до каталога wwwroot
                 uniqueFileName = Guid.NewGuid().ToString() + "_" + photo.FileName; //генерация уникального имени файла
                 string filePath = Path.Combine(uploadsFolder, uniqueFileName); //объединение имени файла и сгенерированного уникального имени
-
+                if (!Directory.Exists(uploadsFolder)) //проверяем существует ли каталог
+                {
+                    Directory.CreateDirectory(uploadsFolder);
+                }
                 //логика сохранения на сервер фото
                 using (var fs = new FileStream(filePath, FileMode.Create))
                 {
